refactor: Update submission card component names and styles
