+++ app/components/features/submission-card/challenge-submission-card.tsx
@@ -5,13 +5,17 @@ import {
   SubmissionCard,
   SubmissionCardFooter,
   SubmissionCardImage,
-} from "./submission-card-v2";
+} from "./submission-card";
 import { Link } from "@remix-run/react";
+import type { Reactions } from "~/lib/models/reactions.server";
+import ReactionsButton from "../reactions/reactions-button";
 
 type ChallengeSubmissionCardProps = {
   submissionImageUrl: string;
   avatar: UserAvatarType;
   challengeSlug?: string;
+  reactions?: Reactions;
+  challengeUserId: number;
   isSolution?: boolean;
 };
 
@@ -19,18 +23,25 @@ export default function ChallengeSubmissionCard({
   submissionImageUrl,
   isSolution = false,
   avatar,
+  reactions,
+  challengeUserId,
   challengeSlug,
 }: ChallengeSubmissionCardProps) {
   const submissionUrl = `/mini-projetos/${challengeSlug}/submissoes/${avatar.github_user}`;
   return (
-    <SubmissionCard size="full">
-      <Link to={submissionUrl}>
+    <SubmissionCard
+      size="full"
+      className="transition-colors dark:hover:border-brand hover:border-brand"
+    >
+      <Link to={submissionUrl} className="bg-black hover:bg-opacity-25">
         <SubmissionCardImage imageUrl={submissionImageUrl} />
       </Link>
       <SubmissionCardFooter>
         <ChallengeSubmissionCardFooter
           isSolution={isSolution}
           avatar={avatar}
+          reactions={reactions}
+          challengeUserId={challengeUserId}
         />
       </SubmissionCardFooter>
     </SubmissionCard>
@@ -40,26 +51,49 @@ export default function ChallengeSubmissionCard({
 function ChallengeSubmissionCardFooter({
   avatar,
   isSolution,
+  reactions,
+  challengeUserId,
 }: {
   avatar: UserAvatarType;
   isSolution: boolean;
+  reactions?: Reactions;
+  challengeUserId: number;
 }) {
   return (
-    <footer className="flex items-center gap-3 ">
-      <UserAvatar avatar={avatar} className="w-10 h-10" />
-      <div className="w-full flex-1">
-        {isSolution ? (
-          <span className="text-brand-500 text-xs">
-            Resolução <b>oficial</b> de
-          </span>
-        ) : (
-          <p className="text-xs dark:text-gray-400">Resolução de</p>
-        )}
-
-        <h3 className=" line-clamp-1" title={formatName(avatar.name)}>
-          {formatName(avatar.name)}
-        </h3>
+    <footer className="flex items-center justify-between gap-2 ">
+      <div className="flex items-center gap-2">
+        <UserAvatar avatar={avatar} className="w-10 h-10" />
+        <UserFooterText isSolution={isSolution} userName={avatar.name} />
       </div>
+      <ReactionsButton
+        reactions={reactions!}
+        reactableId={challengeUserId}
+        reactableType="ChallengeUser"
+      />
     </footer>
+  );
+}
+
+function UserFooterText({
+  isSolution,
+  userName,
+}: {
+  isSolution: boolean;
+  userName: string;
+}) {
+  return (
+    <div className="flex-1 w-full text-xs ">
+      {isSolution ? (
+        <span className="text-xs text-brand-500">
+          Resolução <b>oficial</b> de
+        </span>
+      ) : (
+        <span className="text-xs dark:text-gray-400">Resolução de</span>
+      )}
+
+      <h3 className="text-base line-clamp-1" title={formatName(userName)}>
+        {formatName(userName)}
+      </h3>
+    </div>
   );
 }
