+++ app/routes/_layout-app/_mini-projetos/mini-projetos_.$slug_/_tabs/submissoes/index.tsx
@@ -1,14 +1,14 @@
 import { Link, useOutletContext } from "@remix-run/react";
 import { Send } from "lucide-react";
-import ChallengeSubmissionCard from "~/components/features/submission-card/challenge-submission-card";
-import { Button } from "~/components/ui/button";
 import { Card, CardContent, CardFooter } from "~/components/ui/cards/card";
+import { Button } from "~/components/ui/button";
 import UserAvatar from "~/components/ui/user-avatar";
 import type { Challenge } from "~/lib/models/challenge.server";
 import type { ChallengeUser, User } from "~/lib/models/user.server";
 import classNames from "~/lib/utils/class-names";
 import { formatName } from "~/lib/utils/format-name";
 import WaitingIcon from "../../components/waiting-icon";
+import ChallengeSubmissionCard from "~/components/features/submission-card/challenge-submission-card";
 
 export default function Submissions() {
   const { challengeUsers, challenge, user } = useOutletContext<{
@@ -21,7 +21,7 @@ export default function Submissions() {
 
   return (
     <>
-      <div className="container grid justify-center xl:gap-10 gap-6 lg:grid-cols-3 md:grid-cols-2">
+      <div className="container grid justify-center gap-6 xl:gap-10 lg:grid-cols-3 md:grid-cols-2">
         {/* If user joined challenge */}
         {userStatus === "joined" ||
         userStatus === "forked" ||
@@ -32,15 +32,11 @@ export default function Submissions() {
           <ChallengeSubmissionCard
             submissionImageUrl={challengeUser.submission_image_url}
             avatar={challengeUser.user.avatar}
-            k
-            challengeSlug="desafio-1"
-            // listed={challengeUser.listed}
+            challengeSlug={challenge.slug}
             isSolution={challengeUser.is_solution}
             key={challengeUser.id}
-
-            // challengeUser={challengeUser}
-            // size="md"
-            // challengeSlug={challenge.slug}
+            challengeUserId={challengeUser.id}
+            reactions={challengeUser.reactions}
           />
         ))}
       </div>
@@ -56,16 +52,16 @@ function WaitingSubmissionBanner({
   challenge: Challenge;
 }) {
   return (
-    <Card className="flex flex-col justify-between max-w-[377px]">
-      <CardContent className="flex flex-1 pt-6 items-center justify-center bg-gray-100 dark:bg-transparent">
-        <div className="flex-1">
+    <Card className="flex flex-col justify-between ">
+      <CardContent className="flex items-center flex-1 gap-4 pt-6 bg-gray-100 md:justify-center dark:bg-transparent">
+        <div className="md:flex-1">
           <WaitingIcon />
         </div>
         <div className="w-3/5">
           <h3 className="font-extrabold text-gray-600 dark:text-gray-400">
             Psiu...
           </h3>
-          <p className="text-sm text-gray-600 dark:text-gray-400 mt-2">
+          <p className="mt-2 text-sm text-gray-600 dark:text-gray-400">
             Estamos aguardando a sua submissão 👀
           </p>
           <Button asChild size="sm" className="mt-4">
@@ -82,10 +78,10 @@ function WaitingSubmissionBanner({
             "flex items-center justify-between gap-4 dark:bg-background-700 p-4 w-full",
           )}
         >
-          <div className="w-10 h-10 flex-none">
+          <div className="flex-none w-10 h-10">
             <UserAvatar
               avatar={user.avatar}
-              className="w-10 h-10 flex-shrink-0"
+              className="flex-shrink-0 w-10 h-10"
             />
           </div>
           <div className="w-full">
@@ -93,7 +89,7 @@ function WaitingSubmissionBanner({
               <span>Resolução de</span>
             </h4>
             <h3
-              className="font-semibold line-clamp-1 text-gray-700 dark:text-white"
+              className="font-semibold text-gray-700 line-clamp-1 dark:text-white"
               title={formatName(user.name)}
             >
               {formatName(user.name)}
